+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:rybbit_flutter_sdk/rybbit_flutter.dart';
+import 'package:rybbit_flutter_sdk/rybbit_flutter_sdk.dart';
 import 'package:rybbit_flutter_sdk/ga4_events.dart';
 
 void main() {